+++ scripts/lint.py
@@ -7,11 +7,7 @@ import click
 
 
 def lint():
-    commands = [
-        'isort -c',
-        'flake8',
-        'mypy termlog'
-        ]
+    commands = ["isort -c", "flake8", "mypy termlog"]
     for command in commands:
         try:
             run(shlex.split(command), check=True)
@@ -19,11 +15,12 @@ def lint():
             exit(1)
     else:
         if not commands:
-            print('ERROR: No files found. Run build first')
+            print("ERROR: No files found. Run build first")
             exit(1)
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
+
     @click.command()
     def cli():
         lint()