+++ termlog/config.py
@@ -1,4 +1,4 @@
-from dataclasses import dataclass
+from dataclasses import asdict, dataclass
 from typing import Optional
 
 
@@ -13,21 +13,27 @@ class TerminalConfig:
         time_format: control how the timestamp appears in the output
 
     """
+
     color: bool = True
     json: bool = False
+    time_format: str = "%Y%m%d%H%M%S"
     timestamp: Optional[bool] = None
-    time_format: str = '%Y%m%d%H%M%S'
+
+    def __getitem__(self, item):
+        data = asdict(self)
+        return data[item]
+
+    def keys(self):
+        data = asdict(self)
+        return data.keys()
 
 
 _terminal_config = TerminalConfig()
 
 
 def set_config(
-        color: Optional[bool] = None,
-        json: Optional[bool] = None,
-        time_format: Optional[str] = None,
-        timestamp: Optional[bool] = None,
-        ) -> TerminalConfig:
+    color: Optional[bool] = None, json: Optional[bool] = None, time_format: Optional[str] = None, timestamp: Optional[bool] = None,
+) -> TerminalConfig:
     """Sets configuration for subsequent termlog API calls.
 
     Note:  Generally, True should only be set for color or json
@@ -43,8 +49,12 @@ def set_config(
 
     """
     global _terminal_config
-    _terminal_config.color = color if color is not None else _terminal_config.color
-    _terminal_config.json = json if json is not None else _terminal_config.json
-    _terminal_config.timestamp = timestamp if timestamp is not None else _terminal_config.timestamp
-    _terminal_config.time_format = time_format if time_format is not None else _terminal_config.time_format
+    if color is not None:
+        _terminal_config.color = color
+    if json is not None:
+        _terminal_config.json = json
+    if time_format is not None:
+        _terminal_config.time_format = time_format
+    if timestamp is not None:
+        _terminal_config.timestamp = timestamp
     return _terminal_config