+++ termlog/terminal.py
@@ -6,20 +6,21 @@ from pygments.lexer import Lexer
 from .config import _terminal_config, set_config
 from .formatting import format
 
-__all__ = ('echo', 'set_config')
+__all__ = ("echo", "set_config")
 
 
-def echo(*messages: Any,
-         verbose: Optional[Union[bool, int]] = None,
-         end: str = '\n',
-         flush: bool = True,
-         lexer: Optional[Union[Lexer, str]] = None,
-         file: IO = sys.stdout,
-         color: Optional[bool] = None,
-         json: Optional[bool] = None,
-         time_format: Optional[str] = None,
-         add_timestamp: Optional[bool] = None,
-         ) -> Optional[str]:
+def echo(
+    *messages: Any,
+    verbose: Optional[Union[bool, int]] = None,
+    end: str = "\n",
+    flush: bool = True,
+    lexer: Optional[Union[Lexer, str]] = None,
+    file: IO = sys.stdout,
+    color: Optional[bool] = None,
+    json: Optional[bool] = None,
+    time_format: Optional[str] = None,
+    add_timestamp: Optional[bool] = None,
+) -> Optional[str]:
     """Echo *message*.
 
     lexers can be passed in, and if they are, the resulting output